requirements.txt
python-telegram-bot==20.8
geopy
nest_asyncio
